Extracted and moved processing and hydration logic from Manager\TransactionManager to Transaction\TransactionProcessor, Transaction\TransactionHydrator and Transaction\AuditTrait
